+++ Refureku/Library/Source/TypeInfo/Fields/Field.cpp
@@ -2,7 +2,7 @@
 
 using namespace rfk;
 
-Field::Field(std::string&& name, uint64 id, EFieldFlags flags, Struct const* ownerStruct, Struct const* introducedBy, uint32 memoryOffset, bool isMutable) noexcept:
+Field::Field(std::string&& name, uint64 id, EFieldFlags flags, Struct const* ownerStruct, Struct const* introducedBy, uint64 memoryOffset, bool isMutable) noexcept:
 	FieldBase(std::forward<std::string>(name), id, flags, ownerStruct, introducedBy),
 	memoryOffset{memoryOffset},
 	isMutable{isMutable}
